% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{add_question_answer_col}
\alias{add_question_answer_col}
\title{Create Combined Question and Answer Column}
\usage{
add_question_answer_col(df)
}
\arguments{
\item{df}{Input data frame (tibble) with 'Question' and 'Answer' columns.}
}
\value{
A tibble with the added 'Question And Answer' column.
}
\description{
Concatenates the 'Question' and 'Answer' columns into a single
'Question And Answer' column.
}
