% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeopardy_eda.R
\name{calculate_word_counts}
\alias{calculate_word_counts}
\title{Calculate Word Counts}
\usage{
calculate_word_counts(df)
}
\arguments{
\item{df}{Input data frame (tibble) containing 'Clue Difficulty', 'Answer',
and 'Question' columns. This data should typically be the result of
processing the raw data using `create_regular_episodes_df` with
`add_difficulty = TRUE`.}
}
\value{
A tibble with 'Clue Difficulty', 'Answer Word Count', and
  'Question Word Count' columns.
}
\description{
Calculates the number of words in the 'Answer' and 'Question' columns.
}
