% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaner.R
\name{extract_text_vector}
\alias{extract_text_vector}
\title{Extract Text Vector from DataFrame Column}
\usage{
extract_text_vector(df, col = "Question And Answer")
}
\arguments{
\item{df}{The input data frame (tibble).}

\item{col}{The name of the column to extract (string).}
}
\value{
A character vector containing the text data from the specified column.
}
\description{
Extracts a specific column from a data frame and returns it as a character vector.
}
