% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_model.R
\name{get_nmf_residuals}
\alias{get_nmf_residuals}
\title{Get NMF Residuals}
\usage{
get_nmf_residuals(nmf_result)
}
\arguments{
\item{nmf_result}{An object of class `NMF` (output from `run_nmf_model` or `NMF::nmf`).}
}
\value{
The Frobenius norm of the residuals.
}
\description{
Extracts the Frobenius norm of the residuals (||X - WH||) from a fitted NMF model object.
This serves as a proxy for reconstruction error.
}
