% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_model.R
\name{get_top_words_per_topic}
\alias{get_top_words_per_topic}
\title{Get Top Words Per Topic}
\usage{
get_top_words_per_topic(nmf_result, n_top_words = 10)
}
\arguments{
\item{nmf_result}{An object of class `NMF`.}

\item{n_top_words}{The number of top words to retrieve for each topic.}
}
\value{
A list where each element corresponds to a topic and contains a
  character vector of the top `n_top_words` for that topic.
}
\description{
Extracts the top N most heavily weighted words for each topic from the
basis matrix (W) of a fitted NMF model.
}
