% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeopardy_eda.R
\name{plot_top_categories}
\alias{plot_top_categories}
\title{Plot Top N Categories}
\usage{
plot_top_categories(
  top_cats_df,
  bar_color = "midnightblue",
  save = FALSE,
  output_dir = NULL,
  filename = NULL
)
}
\arguments{
\item{top_cats_df}{Data frame output from `calculate_top_categories`.}

\item{bar_color}{Color for the bars.}

\item{save}{Logical. If TRUE, saves the plot to a file.}

\item{output_dir}{Directory path to save the plot if `save = TRUE`.}

\item{filename}{Name for the saved file.}
}
\value{
Invisibly returns the ggplot object. Displays or saves the plot.
}
\description{
Creates a bar chart showing the top N J-Categories based on their
estimated episode appearances.
}
